test(web): move route tag coverage into middleware tests
